started to separate the methods of the handler class into a DB operations  and a handler classes
